Refactor js-api.d.ts into multiple files
