+++ spec/js-api/legacy/exception.d.ts
@@ -1,0 +1,8 @@
+export interface SassException extends Error {
+  message: string;
+  formatted: string;
+  line: number;
+  column: number;
+  status: number;
+  file: string;
+}
